pyproject.toml
[project]
name = "pdf-scan"
version = "0.1.0"
description = "PDF scanning service for detecting sensitive information"
readme = "README.md"
authors = [
    { name = "Ashkan Aghdai", email = "ashkan.aghdai+2@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "asynch>=0.2.3",
    "fastapi>=0.120.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[project.scripts]
pdf-scan = "pdf_scan:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
